Replace Copy Layer to Layer with the enhanced version
Removes the old, simpler Copy Layer to Layer script and promotes
Copy Layer to Layer (Enhanced) to take its place under the original
name. Updates the README description to reflect the added
capabilities (append mode, background/color layers, font-wide apply,
creating missing target layers).
